clean unnecessary dependencies,make all code works well after upgrade dependencies

+++ AboutIO/src/main/java/NetworkPrograming/Netty/Serializable/ServerHandler.java
@@ -1,7 +1,5 @@
 package NetworkPrograming.Netty.Serializable;
 
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.Unpooled;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 
@@ -15,26 +13,39 @@ import java.util.concurrent.atomic.AtomicInteger;
  */
 public class ServerHandler extends ChannelInboundHandlerAdapter {
 	
-	private AtomicInteger index = new AtomicInteger(0);
+	private final AtomicInteger index = new AtomicInteger(0);
 	
 	@Override
 	public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-		
+		System.out.println("ServerHandler.channelRead() called!");
+
 		System.out.println(msg);
 		Request req = (Request) msg;
 		System.out.println("Server -> " + req.getId() + ", " + req.getMessage());
 		
 		byte[] attachment = GzipUtil.ungzip(req.getAttachment());
 
-		String path = System.getProperty("user.dir") + File.separatorChar + "sources" +
-				File.separatorChar + index.getAndIncrement() + ".jpg";
-
+		// 使用相对路径，在当前工作目录下创建 server_received 目录
+		String relativePath = "AboutIO/src/main/resources/server_received";
+		File directory = new File(relativePath);
+		if (!directory.exists()) {
+			boolean mkdir = directory.mkdir();
+			if (!mkdir) System.err.println("Failed to create directory!");
+			return;
+		}
+		int no = index.getAndIncrement();
+		String path = relativePath + File.separatorChar + no + ".jpg";
 		FileOutputStream fos = new FileOutputStream(path);
 		fos.write(attachment);
 		fos.close();
 		
-		Request res = new Request("1","message1");
+		Request res = new Request(String.valueOf(no),"message from server");
 		ctx.writeAndFlush(res);
 	}
-	
+
+	@Override
+	public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
+		System.err.println("Exception caught in ServerHandler: " + cause.getMessage());
+		ctx.close();
+	}
 }
